third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/writing-mode-parsing-svg1-002-expected.txt

This is a testharness.js-based test.
PASS -webkit-writing-mode: lr-tb 
FAIL -webkit-writing-mode: rl-tb assert_equals: expected "rl-tb" but got "lr-tb"
FAIL -webkit-writing-mode: tb-rl assert_equals: expected "tb-rl" but got "lr-tb"
Harness: the test ran to completion.
